a-star/requirements.txt
numpy 
opencv-python-headless 
pyastar2d 
kafka-python